pyproject.toml
[project]
name = "gooddata-cli"
version = "0.1.0"
description = "CLI tools for interacting with GoodData via Python SDK"
requires-python = ">=3.10"
dependencies = [
    "gooddata-sdk>=1.20.0",
    "gooddata-pandas>=1.20.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
gooddata = "gooddata_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gooddata_cli"]

[tool.ruff]
line-length = 100
target-version = "py310"
